schema: add ModelSchema
